pass a managed instance for save instead of the detached one
Saving the detached object results in it being persisted as a new entity, but it is not.
This is why save() should receive a reference to the managed copy of the object once it has been merged.
If not, then a duplicate object would be needlessly created instead of updating the previously persisted one.